+++ Chaos/Networking/ChaosWorldClient.cs
@@ -52,7 +52,7 @@ public sealed class ChaosWorldClient : WorldClientBase, IChaosWorldClient
     public uint LoginId1 { get; set; }
 
     /// <inheritdoc />
-    public ushort LoginId2 { get; set; }
+    public uint LoginId2 { get; set; }
 
     public ChaosWorldClient(
         Socket socket,
